move helper func to keygen_test.go
