+++ exp/services/recoverysigner/internal/crypto/service_key_insecure_test.go
@@ -3,10 +3,6 @@ package crypto
 import (
 	"testing"
 
-	"github.com/golang/protobuf/proto"
-	"github.com/google/tink/go/hybrid"
-	"github.com/google/tink/go/insecurecleartextkeyset"
-	"github.com/google/tink/go/keyset"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -41,18 +37,4 @@ func TestInsecureServiceKey_encryptDecrypt(t *testing.T) {
 	// context info not matching will result in a failed decryption
 	_, err = serviceKey.Decrypt(ciphertext, []byte("wrong info"))
 	assert.Error(t, err)
-}
-
-func generateHybridEncryptionPrivateKeyKeyset(t *testing.T) []byte {
-	khPriv, err := keyset.NewHandle(hybrid.ECIESHKDFAES128CTRHMACSHA256KeyTemplate())
-	require.NoError(t, err)
-
-	exportedPriv := &keyset.MemReaderWriter{}
-	err = insecurecleartextkeyset.Write(khPriv, exportedPriv)
-	require.NoError(t, err)
-
-	ksPriv, err := proto.Marshal(exportedPriv.Keyset)
-	require.NoError(t, err)
-
-	return ksPriv
 }
